♻️ refactor: index.ts로 컴포넌트 관리
